+++ src/bugit_v2/app.py
@@ -23,9 +23,16 @@ from bugit_v2.bug_report_submitters.mock_jira import MockJiraSubmitter
 from bugit_v2.bug_report_submitters.mock_lp import MockLaunchpadSubmitter
 from bugit_v2.checkbox_utils import Session, get_checkbox_version
 from bugit_v2.models.app_args import AppArgs
-from bugit_v2.models.bug_report import BugReport, PartialBugReport
+from bugit_v2.models.bug_report import (
+    BugReport,
+    PartialBugReport,
+    recover_from_autosave,
+)
 from bugit_v2.screens.bug_report_screen import BugReportScreen
 from bugit_v2.screens.job_selection_screen import JobSelectionScreen
+from bugit_v2.screens.recover_from_autosave_screen import (
+    RecoverFromAutoSaveScreen,
+)
 from bugit_v2.screens.reopen_bug_editor_screen import ReopenBugEditorScreen
 from bugit_v2.screens.reopen_precheck_screen import ReopenPreCheckScreen
 from bugit_v2.screens.session_selection_screen import SessionSelectionScreen
@@ -34,7 +41,7 @@ from bugit_v2.screens.submission_progress_screen import (
     SubmissionProgressScreen,
 )
 from bugit_v2.utils import is_prod, is_snap
-from bugit_v2.utils.constants import NullSelection
+from bugit_v2.utils.constants import AUTOSAVE_DIR, NullSelection
 from bugit_v2.utils.validations import before_entry_check, is_cid
 
 cli_app = typer.Typer(
@@ -84,27 +91,38 @@ def assignee_str_check(value: str | None) -> str | None:
 
 
 @dataclass(slots=True, frozen=True)
-class AppState:
+class NavigationState:
     """
-    The global app state. Check the watch_state function to see all possible
+    The global state for navigation. Check the watch_state function to see all possible
     state combinations
     """
 
     session: Session | Literal[NullSelection.NO_SESSION] | None = None
     job_id: str | Literal[NullSelection.NO_JOB] | None = None
-    bug_report: BugReport | PartialBugReport | None = None
+    bug_report_init_state: (
+        BugReport | Literal[NullSelection.NO_BACKUP] | None
+    ) = None
+    bug_report_to_submit: BugReport | None = None
+
+
+@dataclass(slots=True, frozen=True)
+class ReopenNavigationState:
+    session: Session | Literal[NullSelection.NO_SESSION] | None = None
+    job_id: str | Literal[NullSelection.NO_JOB] | None = None
+    bug_report_to_submit: PartialBugReport | None = None
+    bug_report_init_state: PartialBugReport | None = None
 
 
 @final
 class BugitApp(App[None]):
-    state = var(AppState())
+    nav_state = var[NavigationState | ReopenNavigationState](NavigationState())
     args: AppArgs
     # Any doesn't matter here
     submitter_class: type[
         BugReportSubmitter[Any, Any]  # pyright: ignore[reportExplicitAny]
     ]
-    bug_report_backup: BugReport | None = None
-    partial_bug_report_backup: PartialBugReport | None = None
+    # bug_report_backup: BugReport | None = None
+    partial_bug_report_to_submit_backup: PartialBugReport | None = None
     BINDINGS = [Binding("alt+left", "go_back", "Go Back")]
 
     def __init__(
@@ -173,11 +191,12 @@ class BugitApp(App[None]):
             super()._handle_exception(error)
 
     @work
-    async def watch_state(self) -> None:
+    async def watch_nav_state(self) -> None:
         """Push different screens based on the state"""
+        print(self.nav_state)
 
-        def _write_state(new_state: AppState):
-            self.state = new_state
+        def _write_state(new_state: NavigationState):
+            self.nav_state = new_state
 
         def _pick_editor(
             session: Session | Literal[NullSelection.NO_SESSION],
@@ -188,20 +207,20 @@ class BugitApp(App[None]):
                     session,
                     job_id,
                     self.args,
-                    self.partial_bug_report_backup,
+                    self.partial_bug_report_to_submit_backup,
                 )
                 if self.args.bug_to_reopen
                 else BugReportScreen(
                     session,
                     job_id,
                     self.args,
-                    self.bug_report_backup,
+                    # self.bug_report_to_submit_backup,
                 )
             )
 
-        match self.state:
-            case AppState(session=None, job_id=None, bug_report=None):
-                # init, nothing has been selected yet
+        match self.nav_state:
+            case ReopenNavigationState():
+                # not used right now
                 if (
                     b := self.args.bug_to_reopen
                 ) is not None and not ReopenPreCheckScreen.already_checked:
@@ -216,14 +235,51 @@ class BugitApp(App[None]):
                             msg += f"Error is: {repr(check_result)}"
                         self.exit(return_code=1, message=msg)
 
+            case NavigationState(
+                session=None,
+                job_id=None,
+                bug_report_to_submit=None,
+                bug_report_init_state=None,
+            ):
+                # show the recovery screen if there's no backup
+                recovery_file = await self.push_screen_wait(
+                    RecoverFromAutoSaveScreen(AUTOSAVE_DIR)
+                )
+                if recovery_file is None:
+                    self.nav_state = NavigationState(
+                        None, None, NullSelection.NO_BACKUP, None
+                    )
+                    return
+
+                recovered_report = recover_from_autosave(recovery_file)
+                # trigger this watcher again to actually go to the editor page
+                self.nav_state = NavigationState(
+                    recovered_report.checkbox_session
+                    or NullSelection.NO_SESSION,
+                    recovery_file.job_id or NullSelection.NO_JOB,
+                    recovered_report,
+                )
+
+            case NavigationState(
+                session=None,
+                job_id=None,
+                bug_report_to_submit=None,
+                bug_report_init_state=NullSelection.NO_BACKUP,
+            ):
+
                 def after_session_select(
                     rv: Path | Literal[NullSelection.NO_SESSION] | None,
                 ):
+                    print(rv)
                     match rv:
                         case Path():
-                            self.state = AppState(Session(rv))
+                            self.nav_state = NavigationState(Session(rv))
                         case NullSelection.NO_SESSION:
-                            self.state = AppState(rv, NullSelection.NO_JOB)
+                            self.nav_state = NavigationState(
+                                rv,
+                                NullSelection.NO_JOB,
+                                NullSelection.NO_BACKUP,
+                            )
                         case None:
                             raise RuntimeError(
                                 "Session selection should not return None"
@@ -232,60 +288,70 @@ class BugitApp(App[None]):
                 self.push_screen(
                     SessionSelectionScreen(), after_session_select
                 )
-            case AppState(
-                session=Session() as session, job_id=None, bug_report=None
+
+            case NavigationState(
+                session=Session() as session,
+                job_id=None,
+                bug_report_to_submit=None,
+                bug_report_init_state=None
+                | NullSelection.NO_BACKUP,  # job selection is only possible without init value
             ):
                 # selected a normal session, should go to job selection
                 self.push_screen(
                     JobSelectionScreen(session),
-                    lambda job_id: _write_state(AppState(session, job_id)),
-                )
-            case AppState(
-                session=NullSelection.NO_SESSION as session,
-                job_id=NullSelection.NO_JOB as job_id,
-                bug_report=None,
-            ):
-                # selected no session, skip to editor with absolutely nothing
-                self.push_screen(
-                    _pick_editor(session, job_id),
-                    lambda bug_report: _write_state(
-                        AppState(
-                            session,
-                            job_id,
-                            bug_report,
+                    lambda job_id: _write_state(
+                        NavigationState(
+                            session, job_id, NullSelection.NO_BACKUP
                         )
                     ),
                 )
-            case AppState(
-                session=Session() as session,
-                job_id=NullSelection.NO_JOB as job_id,
-                bug_report=None,
-            ):
-                # has session, but chose the no job object
-                # skip to editor with session
-                self.push_screen(
-                    _pick_editor(session, job_id),
-                    lambda bug_report: _write_state(
-                        AppState(session, job_id, bug_report)
-                    ),
+
+            case (
+                NavigationState(
+                    session=NullSelection.NO_SESSION as session,
+                    job_id=NullSelection.NO_JOB as job_id,
+                    bug_report_to_submit=None,
+                    bug_report_init_state=BugReport()
+                    | NullSelection.NO_BACKUP as init_state,
                 )
-            case AppState(
-                session=Session() as session,
-                job_id=str() as job_id,
-                bug_report=None,
+                | NavigationState(
+                    session=Session() as session,
+                    job_id=NullSelection.NO_JOB as job_id,
+                    bug_report_to_submit=None,
+                    bug_report_init_state=BugReport()
+                    | NullSelection.NO_BACKUP as init_state,
+                )
+                | NavigationState(
+                    session=Session() as session,
+                    job_id=str() as job_id,
+                    bug_report_to_submit=None,
+                    bug_report_init_state=BugReport()
+                    | NullSelection.NO_BACKUP as init_state,
+                )
             ):
                 # normal case, session and job_id were selected
                 # go to editor with info
                 self.push_screen(
-                    _pick_editor(session, job_id),
-                    lambda bug_report: _write_state(
-                        AppState(session, job_id, bug_report)
+                    BugReportScreen(
+                        session,
+                        job_id,
+                        self.args,
+                        (
+                            None
+                            if init_state == NullSelection.NO_BACKUP
+                            else init_state
+                        ),
+                    ),
+                    lambda bug_report_to_submit: _write_state(
+                        NavigationState(
+                            session, job_id, None, bug_report_to_submit
+                        )
                     ),
                 )
-            case AppState(
+            case NavigationState(
                 session=Session() | NullSelection.NO_SESSION as session,
                 job_id=str() | NullSelection.NO_JOB as job_id,
-                bug_report=BugReport() | PartialBugReport() as br,
+                bug_report_to_submit=BugReport() as report,
             ):
                 # returning from the end of submission screen
                 # handle the button selections
@@ -301,33 +367,28 @@ class BugitApp(App[None]):
                         case "quit":
                             self.exit()
                         case "session":
-                            self.bug_report_backup = None
-                            self.partial_bug_report_backup = None
-                            self.state = AppState()
+                            self.nav_state = NavigationState()
                         case "job" if session != NullSelection.NO_SESSION:
-                            # this is only available when there's a session
-                            self.bug_report_backup = None
-                            self.partial_bug_report_backup = None
-                            self.state = AppState(session)
+                            self.nav_state = NavigationState(
+                                session, None, NullSelection.NO_BACKUP
+                            )
                         case "report_editor":
-                            if isinstance(br, PartialBugReport):
-                                self.partial_bug_report_backup = br
-                            else:
-                                self.bug_report_backup = br
-
-                            self.state = AppState(session, job_id)
+                            # restore the report
+                            self.nav_state = NavigationState(
+                                session, job_id, report
+                            )
                         case _:
                             raise RuntimeError()
 
                 self.push_screen(
                     SubmissionProgressScreen(
-                        br, self.submitter_class(), self.args
+                        report, self.submitter_class(), self.args
                     ),
                     after_submission_finished,
                 )
 
             case _:
-                raise RuntimeError(f"Impossible state: {self.state}")
+                raise RuntimeError(f"Impossible state: {self.nav_state}")
 
     def action_go_back(self) -> None:
         """Handles the `Go Back` button
@@ -335,33 +396,45 @@ class BugitApp(App[None]):
         This function should only reassign (not modify) the state object and
         let textual automatically re-render
         """
-        match self.state:
-            case AppState(session=None):
+        match self.nav_state:
+            case NavigationState(
+                session=None,
+                job_id=None,
+                bug_report_init_state=None,
+                bug_report_to_submit=None,
+            ):
+                # backup selection screen
                 self.notify("Already at the beginning")
-                return
-            case AppState(session=Session(), job_id=None):
-                # returning from job selection
-                self.state = AppState(None, None)
-            case AppState(
+            case NavigationState(session=None):
+                self.nav_state = NavigationState()
+            case NavigationState(
+                bug_report_init_state=BugReport()
+            ) | NavigationState(
                 session=NullSelection.NO_SESSION,
                 job_id=NullSelection.NO_JOB,
-                bug_report=None,
+                bug_report_to_submit=None,
             ):
-                # returning from editor with nothing selected
-                # just go back to the beginning
-                self.state = AppState(None, None)
-            case AppState(
+                # 1. started with a backup
+                # 2. returning from editor with nothing selected
+                # go to backup selection
+                self.nav_state = NavigationState()
+            case NavigationState(session=Session(), job_id=None):
+                # returning from job selection
+                self.nav_state = NavigationState(
+                    None, None, NullSelection.NO_BACKUP, None
+                )
+            case NavigationState(
                 session=Session() as session,
                 job_id=str() | NullSelection.NO_JOB,
-                bug_report=None,
+                bug_report_init_state=NullSelection.NO_BACKUP,
             ):
                 # returning from editor with explicit job selection
                 # go back to job selection
-                self.state = AppState(session, None)
-            case AppState(
+                self.nav_state = NavigationState(session, None)
+            case NavigationState(
                 session=Session() | NullSelection.NO_SESSION,
                 job_id=str() | NullSelection.NO_JOB,
-                bug_report=BugReport() | PartialBugReport(),
+                bug_report_to_submit=BugReport() | PartialBugReport(),
             ):
                 self.notify(
                     title="Cannot go back while a submission is happening",
@@ -369,7 +442,7 @@ class BugitApp(App[None]):
                 )
             case _:
                 raise RuntimeError(
-                    f"Impossible state when going back: {self.state}"
+                    f"Impossible state when going back: {self.nav_state}"
                 )