+++ src/bugit_v2/screens/session_selection_screen.py
@@ -33,7 +33,7 @@ class SessionSelectionScreen(Screen[Path | Literal[NullSelection.NO_SESSION]]):
     def compose(self) -> ComposeResult:
         # textual's header crashes for some reason
         yield SimpleHeader()
-        with Vertical(classes="w100 h100 center", id="after_load_container"):
+        with Vertical(classes="w100 h100 center"):
             yield Label("[b][$primary]Select a Session")
             yield VerticalScroll(
                 Button(
@@ -42,13 +42,15 @@ class SessionSelectionScreen(Screen[Path | Literal[NullSelection.NO_SESSION]]):
                     # tooltip != None is used to check if this special
                     # button is clicked, do not remove
                     tooltip="Choose this to skip to report editor",
-                    classes="mb1 session_button",
+                    classes="session_button",
+                    flat=True,
                 ),
                 *(
                     Button(
                         os.path.basename(session),
                         name=str(session),
-                        classes="mb1 session_button",
+                        classes="session_button",
+                        flat=True,
                     )
                     for session in self.session_dirs
                 ),