+++ CHANGELOG.md
@@ -1,14 +1,26 @@
+# 0.2.5
+
+## Added
+
+- Autosave functionality: Now the editor saves your progress as you type. It utilizes the [debounce](https://medium.com/@jamischarles/what-is-debouncing-2505c0648ff1) pattern to ensure performance while editing and only saves the progress to disk when the user hasn't been typing for 1 second.
+- Recovery from autosave: If autosave files exist on the system, bugit will show a recovery screen asking the user whether they want to use any of the previous saved files.
+- Added a clear button for the log selection list
+
 # 0.2.4
 
 ## Added
 
 - CLI arguments to pre-fill values like CID, SKU, tags, etc. Bugit will hold on to these values and fill them in at the bug report screen. **Note that any human-entered value has higher precedence than CLI values.** So if a submission failed and bugit returned to the editor, the CLI values will be overridden by the values entered by the user.
-- Log collector progress watchers. Now there's a progress message every 30seconds for very slow log collectors
+- Log collector progress watchers. Now there's a progress message every 30 seconds for very slow log collectors
 - Experimental reopen editor implemented, locked until 0.2.5 for now
 - Upgraded to textual 6.1
 - New header style
 - Not specifying any subcommand will now show a help page instead of just an error
-- Fixed the issue of Checkbox version not appearing in the header
+
+## Fixed
+- Checkbox version not appearing in the header
+- Nvidia log collector not getting all the information (lots of missing command or .so files)
+- Missing bug status chooser in the editor when using th launchpad submitted
 
 # 0.2.3